Improve job deletion: Try to remove associated task if job is not running.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@10315 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/appengine/dao/JobDAO.java
@@ -28,10 +28,13 @@ import static com.googlecode.objectify.ObjectifyService.ofy;
 import java.util.List;
 
 import org.sosy_lab.cpachecker.appengine.entity.Job;
+import org.sosy_lab.cpachecker.appengine.entity.Job.Status;
 
+import com.google.appengine.api.taskqueue.Queue;
+import com.google.appengine.api.taskqueue.QueueFactory;
 import com.googlecode.objectify.Key;
 import com.googlecode.objectify.ObjectifyService;
-import com.googlecode.objectify.VoidWork;
+import com.googlecode.objectify.Work;
 
 
 public class JobDAO {
@@ -54,24 +57,53 @@ public class JobDAO {
     return job;
   }
 
-  public static void delete(final Job job) {
-    ofy().transact(new VoidWork() {
+  /**
+   * Tries to delete a job and indicates if it was possible.
+   * Also deletes all associated files.
+   * If the job is in pending state removal of the corresponding task will be tried.
+   * If the job is in running state the job cannot be deleted.
+   *
+   * @param job The job to delete.
+   * @return True, if the job could be deleted, false otherwise.
+   */
+  public static boolean delete(final Job job) {
+    return ofy().transact(new Work<Boolean>() {
 
       @Override
-      public void vrun() {
+      public Boolean run() {
+
+        if (job.getStatus() == Status.PENDING) {
+          try {
+            Queue queue = QueueFactory.getQueue(job.getQueueName());
+            boolean couldDeleteTask = queue.deleteTask(job.getTaskName());
+
+            if (!couldDeleteTask) {
+              return false;
+            }
+          } catch (Exception e) {
+            return false;
+          }
+        }
+
+        if (job.getStatus() == Status.RUNNING) {
+          return false;
+        }
+
         ofy().delete().entities(job.getFiles()).now();
         ofy().delete().entities(job).now();
+
+        return true;
       }
     });
   }
 
-  public static void delete(String key) {
+  public static boolean delete(String key) {
     Key<Job> jobKey = Key.create(key);
-    delete(jobKey);
+    return delete(jobKey);
   }
 
-  public static void delete(Key<Job> key) {
-    delete(load(key));
+  public static boolean delete(Key<Job> key) {
+    return delete(load(key));
   }
 
   public static Key<Job> allocateKey() {
